demo-app/simple-chat/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
azure-search-documents==11.4.0
azure-core==1.29.6
python-dotenv==1.0.0